implement getRecipients function and additional guards
